fix: Add config validation, buffer overflow guards, and error handling (#27)
- Add SimulationConfig::validate() with range checks to prevent NaN/Infinity
  injection into GPU uniforms (9 new tests)
- Widen buffer size calculations to u64 in pipeline.rs and headless.rs to
  prevent u32 overflow at high resolutions
- Use saturating_add for frame_counter
- Clamp scatterer sigma to min 1e-6 to prevent near-zero division
- Log create_dir_all errors instead of silently discarding
- Replace channel send unwrap with let _ pattern
- Slim image crate to PNG-only (removes ~30 unused transitive deps)

Closes #27

+++ src/render/headless.rs
@@ -442,7 +442,7 @@ impl HeadlessRenderPipeline {
         let unpadded_bytes_per_row = self.width * bytes_per_pixel;
         let align = wgpu::COPY_BYTES_PER_ROW_ALIGNMENT;
         let padded_bytes_per_row = unpadded_bytes_per_row.div_ceil(align) * align;
-        let buffer_size = (padded_bytes_per_row * self.height) as u64;
+        let buffer_size = padded_bytes_per_row as u64 * self.height as u64;
 
         // Create staging buffer
         let staging_buffer = self.device.create_buffer(&wgpu::BufferDescriptor {
